refactor: rearrange and rename some functions
